backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyJWT==2.9.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
openpyxl>=3.1.2
alembic==1.13.3
cryptography==43.0.3
httpx==0.27.2
